ananicy-cpp: Package cachyos-ananicy-rules in a separate package
Packaging the rules in a separate package enables us to utilize copr's autobuilds with minimal risk.
It also prevents the need to rebuild ananicy-cpp everytime we want to update the rules we provide.
This way, users can update the rules in their system without the need to update ananicy-cpp itself.

I have also opted to change the package name to match the PKGBUILD. There is no need for manual user
intervention because everything is handled in the .spec file.

Signed-off-by: Eric Naim <dnaim@cachyos.org>