kernel-cachyos{,-lts}: Use openssl-provider patch from upstream kernel-patches repo
For the stable kernel, we're just going to use the patch from 6.6 since it applies there too and the fact that 6.12
is just around the corner and the patch is already upstreamed there.

I've also added some spacing for the NVIDIA stuff to make it a teensy bit more readable.

Removed the ifcond since all versions we build for already has OpenSSl >= 3.0, and it doesn't really hurt to replace
a deprecated API.

Signed-off-by: Eric Naim <dnaim@cachyos.org>