kernel-cachyos/nvidia: Workaround COPR build environment
COPR uses one source rpm for all of the chroots. Since Fedora is not the only RPM-based distribution, rpmbuild needs
to "hardcode" its own internal version. As it turns out, the hardcoded version seems to be F40 or atleast not F41.
This leads it to fetch sources for F40. It will fetch the other missing sources/patches for a specific version, e.g. F41,
during build time. However, since we have the same patch names for %{PATCH4}, it doesn't refetch the version for 565,
causing it to fail to patch.

We can workaround this by renaming the patch in our copr github so it refetches the 565 patch.

Signed-off-by: Eric Naim <dnaim@cachyos.org>