scx-scheds: add bpftool dependency to fix build error and add devel package for devel files
The package failed to build because the command bpftool can't be found. According to the scx Github-page, it's not required to use the latest version or the version, which is being cloned by the compilation script - the system bpftool can also be used.

Devel package has been added for development files, which have to be packaged.