kernel-cachyos: Initial rewrite of the kernel rpmbuild spec
This is an attempt to rewrite the kernel build spec to be much more readable and easier to grasp. Some liberties
were taken during this initial rewrite, namely that it only supports x86_64 for now. This *should* be the
only architecture that we need to support, because the CachyOS patchset is primarily for x86_64.

Firstly, all parts of the build process have proper indentation to make it easier to read. Each declaration
of a variable (name, version, builddeps) are also properly spaced such that everything is aligned. Second,
building kernel-headers was removed. The reason is because that this part of the kernel doesn't need to be built
by us, and it doesn't have to have the same kernel version as the installed system either. This package
is mainly used to build toolchains and core libraries. For some context, see the Arch equivalent package
linux-api-headers[2]

Moved subpackage declarations down the file to join it with their %post and %files counterparts.

I've used the %autopatch macro instead of calling patch manually. This has the added benefit of
keeping the patching stage in one line. Some groundrules need to made here to accomodate this change, namely
ALL source and patch indexes from 0-9 are reserved solely for the kernel. When NVIDIA gets added back, it will
occupy indexes 10-19. Lastly, the generation of symvers is compressed with zstd instead of xz in upstream
Fedora, and instead of gz which was used previously in our spec.

Note that this is the initial rewrite of the spec. It only supports GCC building for now and has no build options
to tune x86_64 ISA levels or the tickrate of the kernel. This will come soon until it has been confirmed
that the spec works correctly.

[1] https://gitlab.archlinux.org/archlinux/packaging/packages/linux/-/blob/main/PKGBUILD
[2] https://gitlab.archlinux.org/archlinux/packaging/packages/linux-api-headers

Signed-off-by: Eric Naim <dnaim@cachyos.org>