kernel-cachyos: Create stub initramfs to aide rpm disk space requirement estimations
The previous size of 48M doesn't seem to be enough for modern systems now. I've collected results from
some users and it seems that the set value of 90M should be a safe middleground, without making it too high.
For reference, an initramfs image with NVIDIA modules bundled can take up to 160M in size.

Signed-off-by: Eric Naim <dnaim@cachyos.org>