kernel-cachyos: Add support for building a minimal kernel
During the packaging process, I've had to rebuild the kernel MULTIPLE times due to stupid typos and other
careless mistakes. This is a very painful process so I bit the bullet and added support for localmodconfig.
This made my build times 2x faster and should be helpful during building in the -dev repositories.

Signed-off-by: Eric Naim <dnaim@cachyos.org>