kernel-cachyos: Adds support for building nvidia-open kernel modules with the kernel
Some concerns of packaging this is that NVIDIA driver versions between different Fedora releases can differ.
We need to clearly define which versions of Fedora we will support. The latest stable is a given. Rawhide is
also a good version to support as it can help us catch some packaging chinks that may need to be changed
in the next version. We do also have some Rawhide users. I'm not a fan of supporting older Fedora versions,
and it only adds complexity to the spec overtime if there are version mismatches.

Signed-off-by: Eric Naim <dnaim@cachyos.org>