kernel-cachyos: Allow customizing the x86_64 microarchitecture ISA level
This is an easy enough thing that can be added to the spec. I have ideas to allow building the kernel with
per-CPU architectures too like ZEN3 or ALDERLAKE, but through real-world benchmarks it is shown that there are
impercievable differences between x86_64_v1 and x86_64_v3 so I'm hesitant on adding more complexity to the spec
for no benefit all.

Signed-off-by: Eric Naim <dnaim@cachyos.org>