kernel-cachyos: Provide the base kernel packages
In Fedoraland, a package is classified as a kernel package if that package provides either `kernel` or
`kernel-core`. kernel-debug-core specifically provides `kernel` to correctly be classified as a kernel package.

Note that this doesn't seem to really matter on Fedora Workstation and equivalents. It is however needed on
OSTree systems, because rpm-ostree matches a package's provides tags to a hardcoded list of strings for kernel
packages.

Provide the other base packages for each kernel package to maintain consistency (e.g. kernel-modules, kernel-devel)

Link: https://github.com/coreos/rpm-ostree/pull/2037
Signed-off-by: Eric Naim <dnaim@cachyos.org>