kernel-cachyos: Move modules.builtin* to kernel-cachyos-modules
There are some depmod warnings currently when installing this kernel. This is due to kenrel-cachyos-modules
owning modules.order while -core owns modules.builtin*. This can be fixed by moving the depmod scriptlet
to posttrans, but issues will arise when depmod is ran in %postun. This reverts things back to how they were
in the old spec until a proper solution is found (maybe we have to create split packages for -modules and
-modules-core).

Signed-off-by: Eric Naim <dnaim@cachyos.org>