kernel-cachyos: Remove indents
Unfortunate as it is, rpm 4.19 which is still used in Fedora 40 doesn't support indents. So for now,
begrudgingly remove all indents. This was done by replacing 4 whitespace characters with a blank character.
The spec probably won't look pretty after this, and I'm not gonna look at it. Thankfully, F40 is EOL soon.

Signed-off-by: Eric Naim <dnaim@cachyos.org>